refactor: PipePluginAgent
